refactor: 刪除 RecipeUploadStep2 元件以簡化結構，並更新引用路徑以改善檔案組織。
